fix: resolve remaining CppUTest StringFrom ambiguity on ARM GCC 15
- Cast Ticks (int64_t) and uint64_t to int32_t in CHECK_EQUAL at lines 520, 810
  of stktest_kernelservice.cpp to fix StringFrom overload ambiguity

+++ test/generic/stktest_kernelservice.cpp
@@ -517,7 +517,7 @@ TEST(KernelService, SleepUntilMissDeadline)
     // task2 calls Sleep to become idle
     SleepUntil(now);
 
-    CHECK_EQUAL(now, GetTicks());
+    CHECK_EQUAL((int32_t)now, (int32_t)GetTicks());
 
     // task2 is still active as it did not sleep
     CHECK_EQUAL_TEXT(active->SP, (size_t)task2.GetStack(), "expecting task2 after sleep");
@@ -807,7 +807,7 @@ TEST(KernelService, SysTimer)
     uint64_t count = stk::GetSysTimerCount();
 
     CHECK_EQUAL(1000, freq);
-    CHECK_EQUAL(99, count);
+    CHECK_EQUAL(99, (int32_t)count);
 }
 
 } // namespace stk
